managment-system3/backend/requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
flask-migrate==4.0.7
flask-bcrypt==1.0.1
flask-jwt-extended==4.6.0
python-dotenv==1.0.1
mysqlclient==2.2.4
marshmallow==3.21.1
marshmallow-sqlalchemy==1.1.0
python-dateutil==2.9.0

